remove the capability to restart the pipeline after a HALT.
In general, transformations which emit more than one element may be
subject to data loss during a halt, so an error-free restart cannot
be guaranteed. It is better to eliminate the possibility altogether,
focus on correctness, and keep the codebase simple. The restart
capability would make all HALT stages much more complicated, so that
correct interoperability between all halt-capable stages is ensured.